Change semantics of MARK-OBJECT for non-semi, non-boundary units
In particular, if a unit moves the supplied cursor "backwards", that
is in the opposite direction of the specified direction, the
MARK-OBJECT operation moves the /point/ cursor "backwards" to that
location in addition the ordinary behavior of moving the mark cursor.

This allows units like LINE to mark the whole lien independent of the
location of the point cursor. In contrast, the SEMI-LINE unit will
result in a region which is delimited by the current location of the
point cursor.