Reorganize {CURSOR,SITE}-CONDITION condition classes
When an operation on a site fails, the site object is not always
available where the error condition is signaled. The new condition
class MAYBE-SITE-CONDITION allows the site object to be attached to
the condition in a place other than the place in which the condition
is signaled. The relevant section of the changed condition hierarchy
looks like this:

EDITING-CONDITION
+-CURSOR-CONDITION
`-SITE-CONDITION
  +-MAYBE-SITE-CONDITION
  `-REQUIRED-SITE-CONDITION

An :AROUND method on PERFORM specialized to the SITE looks at signaled
conditions and attaches the site to conditions of type
MAYBE-SITE-CONDITION if required.